+++ dojo_8/java/src/main/java/com/mercadolibre/dojos/inconsistencies/Inconsitency.java
@@ -1,0 +1,11 @@
+package com.mercadolibre.dojos.inconsistencies;
+
+import com.mercadolibre.dojos.dto.CheckoutOptionsDto;
+
+/**
+ * Created by fdaranno on 20/12/17.
+ */
+public interface Inconsitency {
+    public Integer getNumber();
+    public boolean apply(CheckoutOptionsDto checkoutOptions);
+}
